backend/pyproject.toml
[project]
name = "payout-app"
version = "0.1.0"
description = ""
authors = [
    {name = "Jenoliya",email = "jenoliyasunilkumar@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "djangorestframework (>=3.17.1,<4.0.0)",
    "django-filter (>=25.2,<26.0)",
    "django-oauth-toolkit (>=3.2.0,<4.0.0)",
    "jwt (>=1.4.0,<2.0.0)",
    "celery (>=5.6.3,<6.0.0)",
    "redis (>=7.4.0,<8.0.0)",
    "django-celery-beat (>=2.9.0,<3.0.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
